+++ config/test.exs
@@ -5,13 +5,13 @@ import Config
 # The MIX_TEST_PARTITION environment variable can be used
 # to provide built-in test partitioning in CI environment.
 # Run `mix help test` for more information.
-config :ex_sd, ExSd.Repo,
-  username: "postgres",
-  password: "postgres",
-  hostname: "localhost",
-  database: "ex_sd_test#{System.get_env("MIX_TEST_PARTITION")}",
-  pool: Ecto.Adapters.SQL.Sandbox,
-  pool_size: 10
+# config :ex_sd, ExSd.Repo,
+#   username: "postgres",
+#   password: "postgres",
+#   hostname: "localhost",
+#   database: "ex_sd_test#{System.get_env("MIX_TEST_PARTITION")}",
+#   pool: Ecto.Adapters.SQL.Sandbox,
+#   pool_size: 10
 
 # We don't run a server during test. If one is required,
 # you can enable the server option below.